+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/core/TemplatesToDot.java
@@ -64,6 +64,7 @@ public class TemplatesToDot extends ProvToDot {
 
 
     public static String createHtmlTable(TemplateInfo templateInfo,
+                                         boolean withIcons,
                                          String iconDirectory,
                                          List<String> inputsNames,
                                          List<String> inputsPorts,
@@ -74,21 +75,29 @@ public class TemplatesToDot extends ProvToDot {
         StringBuilder html = new StringBuilder();
 
 
-        //String iconTest="<svg width=\"50\"  height=\"50\"><use href=\"#icon-org.openprovenance.book.fs.FileInit\"/></svg>";
-        //String iconTest0="<use href=\"#icon-org.openprovenance.book.fs.FileInit\"/>";
-        //String iconTest2="<IMG SRC=\"http://localhost:7075/book/webjars/template-intro1/0.1.0-SNAPSHOT/icons/org.openprovenance.book.fs.FileInit.svg\"/>";
-        //String iconTest3="<IMG SRC=\"file:///Users/luc/git-papers/papers/book-ptm/project/template-intro1/src/main/resources/icons/org.openprovenance.book.fs.FileApproving.svg\"/>";
-        String iconImage="<IMG SRC=\"" + iconDirectory + "/" + templateInfo.template + ".png\"/>";
-                // Start building the HTML for the table
-
+        // Start building the HTML for the table
         html.append("<TABLE BORDER=\"0\" CELLBORDER=\"1\" CELLSPACING=\"0\" CELLPADDING=\"4\">\n");
 
         // First row with rowspan and input cells
         html.append("  <TR>\n");
-        html.append(String.format("    <TD ROWSPAN=\"3\" HREF=\"%s\"  TARGET=\"_blank\">%s</TD>\n", templateInfo.url, iconImage));
-        //html.append(String.format("    <TD ROWSPAN=\"3\">%s</TD>\n",  iconImage));
+
+        if (withIcons) {
+            // IMPORTANT NOTE
+            // While graphviz documentation indicates that svg images <IMG> are permitted, in practice it does not seem to be the case.
+            // Thus, we refer to png file.
+            // Such image files MUST be on the file system.
+            // When generating an SVG, graphviz does not embed the images, but instead links to them.
+            // This is problematic, when the visualisation is served by a service to a browser, because the browser will not be able to dereference a file image.
+            // For this to work, the client needs to rewrite the url.
+            // This is implemented by function rewriteImageHrefs(svgElement) in form.html
+            String iconImage="<IMG SRC=\"" + iconDirectory + "/" + templateInfo.template + ".png\"/>";
+            html.append(String.format("    <TD ROWSPAN=\"3\" HREF=\"%s\"  TARGET=\"_blank\">%s</TD>\n", templateInfo.url, iconImage));
+        }
 
         html.append(String.format("    <TD ROWSPAN=\"3\" HREF=\"%s\"  TARGET=\"_blank\">%s </TD>\n", templateInfo.url, templateInfo.templateId));
+        if (inputsNames.isEmpty()) {
+            html.append("    <TD></TD>\n");
+        }
         for (int i = 0; i < inputsNames.size(); i++) {
             html.append(String.format("    <TD PORT=\"%s\" BGCOLOR=\"%s\" HREF=\"%s\" TARGET=\"_blank\">%s</TD>\n",
                     inputsPorts.get(i), inputsColors.get(i), templateInfo.url.replace(".svg", "/"+inputsNames.get(i)), inputsNames.get(i)));
@@ -105,6 +114,8 @@ public class TemplatesToDot extends ProvToDot {
 
         // Close the table
         html.append("</TABLE>");
+
+       // System.out.println(html.toString());
 
         return html.toString();
     }
@@ -267,7 +278,7 @@ public class TemplatesToDot extends ProvToDot {
             List<String> outputsColors = outputsNames.stream().map(s -> provcolors.get(templateBaseTypes.get(s))).collect(Collectors.toList()); //outputsPorts.stream().map(s -> "orange").collect(Collectors.toList());
 
 
-            String html = createHtmlTable(templateInfo, iconDirectory, inputsNames, inputPorts, inputsColors, outputsNames, outputsPorts, outputsColors);
+            String html = createHtmlTable(templateInfo, withIcons, iconDirectory, inputsNames, inputPorts, inputsColors, outputsNames, outputsPorts, outputsColors);
 
             emitTemplate(template, templateId, html, out);